pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "joblib>=1.5.2",
    "loguru>=0.7.3",
    "pip>=25.2",
    "pyspellchecker>=0.8.3",
    "scikit-learn>=1.7.2",
    "spacy>=3.8.7",
    "uvicorn>=0.36.0",
]
